requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
mysql-connector-python==9.5.0
python-dotenv==1.2.1
httpx==0.24.1
requests==2.32.5
